homework_06/pyproject.toml
[tool.poetry]
name = "homework_06"
version = "0.1.0"
description = ""
authors = ["Inna Sorina <sorina-inna@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.1.1"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
pg8000 = "^1.24.1"
Flask-WTF = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
